pyproject.toml
[project]
name = "embedding-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.0",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
markers = [
    "ut: unit test",
    "it: integration test"
]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/__init__.py"
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "testcontainers>=4.10.0",
]
